Breaking up util: Move GinCmd to shell subpackage
Renaming GinCmd to Cmd.
