+++ git/shell/cmd.go
@@ -1,4 +1,5 @@
-package util
+// Package shell augments the builtin os/exec Cmd struct and functions with convenience functions for reading piped output.
+package shell
 
 import (
 	"bufio"
@@ -6,9 +7,9 @@ import (
 	"os/exec"
 )
 
-// GinCmd extends the exec.Cmd struct with convenience functions for reading piped output.
+// Cmd extends the exec.Cmd struct with convenience functions for reading piped output.
 // It also overrides the Wait() method such that the error is stored in the Err field.
-type GinCmd struct {
+type Cmd struct {
 	*exec.Cmd
 	OutReader *bufio.Reader
 	ErrReader *bufio.Reader
@@ -16,17 +17,17 @@ type GinCmd struct {
 }
 
 // Command returns the GinCmd struct to execute the named program with the given arguments.
-func Command(name string, args ...string) GinCmd {
+func Command(name string, args ...string) Cmd {
 	cmd := exec.Command(name, args...)
 	outpipe, _ := cmd.StdoutPipe()
 	errpipe, _ := cmd.StderrPipe()
 	outreader := bufio.NewReader(outpipe)
 	errreader := bufio.NewReader(errpipe)
-	return GinCmd{cmd, outreader, errreader, nil}
+	return Cmd{cmd, outreader, errreader, nil}
 }
 
 // OutputError runs the command and returns the standard output and standard error as two byte slices.
-func (cmd *GinCmd) OutputError() ([]byte, []byte, error) {
+func (cmd *Cmd) OutputError() ([]byte, []byte, error) {
 	var bout, berr bytes.Buffer
 	cmd.Stdout = &bout
 	cmd.Stderr = &berr
@@ -35,7 +36,7 @@ func (cmd *GinCmd) OutputError() ([]byte, []byte, error) {
 }
 
 // Output runs the command and returns its standard output.
-func (cmd *GinCmd) Output() ([]byte, error) {
+func (cmd *Cmd) Output() ([]byte, error) {
 	cmd.Stdout = nil
 	return cmd.Cmd.Output()
 }
